Will this work?
huh?

+++ Github Robot Final/correctPosition.java
@@ -3,7 +3,7 @@ import rxtxrobot.RXTXRobot;
 
 
 public class correctPosition {
-	public void correctPosition(RXTXRobot r, double[] Changes){
+	public void correctPosition(RXTXRobot r, int[] Changes){
 		EForward FEncoded= new EForward();
 		EBackward BEncoded=new EBackward();
 		ELeft LEncoded=new ELeft();
@@ -11,7 +11,7 @@ public class correctPosition {
 		
 		if (Changes[0]>0){
 			REncoded.ERight(r, 100);//turns right
-			FEncoded.EForward(r,(int) (Changes[0]*30) );//moves forward to new position
+			FEncoded.EForward(r, (Changes[0]*30) );//moves forward to new position
 			LEncoded.ELeft(r, 100);//turns left again
 			
 		}
@@ -21,7 +21,7 @@ public class correctPosition {
 			Changes[0]=(Changes[0]*(-1));//Changed to + to interact with run motor
 			
 			LEncoded.ELeft(r, 100);//turns left
-			FEncoded.EForward(r,(int) (Changes[0]*30) );//moves forward to new position
+			FEncoded.EForward(r,(Changes[0]*30) );//moves forward to new position
 			REncoded.ERight(r, 100);//turns right again
 			
 			Changes[0]=(Changes[0]*(-1));//Changed back
@@ -29,12 +29,12 @@ public class correctPosition {
 		}
 	
 		if(Changes[1]>0){
-			r.runMotor(RXTXRobot.MOTOR1,- 100, RXTXRobot.MOTOR2, 130,(int) (Changes[0]*30));//move forward to new position
+			r.runMotor(RXTXRobot.MOTOR1,- 100, RXTXRobot.MOTOR2, 130,(Changes[0]*30));//move forward to new position
 		}
 		
 		if(Changes[1]<0){
 			Changes[1]=(Changes[1]*(-1));//Changed to + to interact with run motor
-			r.runMotor(RXTXRobot.MOTOR1,100, RXTXRobot.MOTOR2, -130,(int) (Changes[0]*30));//move backward to new position
+			r.runMotor(RXTXRobot.MOTOR1,100, RXTXRobot.MOTOR2, -130, (Changes[0]*30));//move backward to new position
 			Changes[1]=(Changes[1]*(-1));//Changed back
 					
 		}
